requirements.txt
streamlit
pandas
easyocr
opencv-python-headless
pillow
numpy